### 2023-03-10 #5243 @editorialbot
[PRE REVIEW]: Sarracen: a Python package for analysis and visualization of smoothed particle hydrodynamics data
**Submitting author:** @ttricco (<a href="http://orcid.org/0000-0002-6238-9096">Terrence Tricco</a>)
**Repository:** https://github.com/ttricco/sarracen
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
