### 2023-04-04 4150afe
Expand documentation for SarracenDataFrame construction from numpy array or dicts
