### 2023-04-10 ba95f2f
Fix pixel culling for exact interpolation
