### 2023-04-10 94bdc3e
Gradient unit test for exact interpolation particle culling
