### 2023-04-12 995f151
Simplified unit test for exact interpolation cell culling
