### 2023-04-12 a6755c2
Revert floor/ceiling in fast interpolation functions
