### 2023-04-17 a7a714c
Add option to impose a minimum smoothing length based on grid size
