+++ sarracen/interpolate/interpolate.py
@@ -341,11 +341,24 @@ def _get_weight(data: 'SarracenDataFrame', target: Union[str, np.ndarray], dens_
         rho_data = _get_density(data)
         return target_data * mass_data / rho_data
 
+def _get_smoothing_lengths(data: 'SarracenDataFrame', hmin: float, x_pixels: int, y_pixels: int,
+                           xlim: Tuple[float, float], ylim: Tuple[float, float]):
+    """ Return the smoothing length data, imposing a minimum length if hmin is True. """
+    
+    if hmin:
+        pix_size = (xlim[1] - xlim[0]) / x_pixels
+        pix_size = np.maximum(pix_size, (ylim[1] - ylim[0]) / y_pixels)
+        h_data = np.maximum(data[data._hcol].to_numpy(), 0.5 * pix_size)
+    else:
+        h_data = data[data._hcol].to_numpy()
+
+    return h_data
+
 
 def interpolate_2d(data: 'SarracenDataFrame', target: str, x: str = None, y: str = None, kernel: BaseKernel = None,
                    x_pixels: int = None, y_pixels: int = None, xlim: Tuple[float, float] = None,
-                   ylim: Tuple[float, float] = None, exact: bool = False,
-                   backend: str = None, dens_weight: bool = False, normalize: bool = True) -> np.ndarray:
+                   ylim: Tuple[float, float] = None, exact: bool = False, backend: str = None,
+                   dens_weight: bool = False, normalize: bool = True, hmin: bool = False) -> np.ndarray:
     """
     Interpolate particle data across two directional axes to a 2D grid of pixels.
 
@@ -375,6 +388,9 @@ def interpolate_2d(data: 'SarracenDataFrame', target: str, x: str = None, y: str
         The computation backend to use when interpolating this data. Defaults to the backend specified in `data`.
     dens_weight: bool
         If True, the target will be multiplied by density. Defaults to False.
+    hmin: bool
+        If True, a minimum smoothing length of 0.5 * pixel size will be imposed. This ensures each particle
+        contributes to at least one grid cell / pixel. Defaults to False (this may change in a future verison).
 
     Returns
     -------
@@ -404,14 +420,16 @@ def interpolate_2d(data: 'SarracenDataFrame', target: str, x: str = None, y: str
     kernel = kernel if kernel is not None else data.kernel
     backend = backend if backend is not None else data.backend
 
+    h_data = _get_smoothing_lengths(data, hmin, x_pixels, y_pixels, xlim, ylim)
+
     grid = get_backend(backend). \
-           interpolate_2d_render(data[x].to_numpy(), data[y].to_numpy(), w_data, data[data._hcol].to_numpy(), kernel.w,
+           interpolate_2d_render(data[x].to_numpy(), data[y].to_numpy(), w_data, h_data, kernel.w,
                               kernel.get_radius(), x_pixels, y_pixels, xlim[0], xlim[1], ylim[0], ylim[1], exact)
 
     if normalize:
         w_norm = _get_weight(data, np.array([1] * len(w_data)), dens_weight)
         norm_grid = get_backend(backend). \
-                    interpolate_2d_render(data[x].to_numpy(), data[y].to_numpy(), w_norm, data[data._hcol].to_numpy(),
+                    interpolate_2d_render(data[x].to_numpy(), data[y].to_numpy(), w_norm, h_data,
                                           kernel.w, kernel.get_radius(), x_pixels, y_pixels, xlim[0], xlim[1], ylim[0],
                                           ylim[1], exact)
         grid = np.nan_to_num(grid / norm_grid)
@@ -422,7 +440,7 @@ def interpolate_2d(data: 'SarracenDataFrame', target: str, x: str = None, y: str
[…]
