### 2023-04-17 #5263 @JBorrow
@ttricco the only thing missing for me now is a clear set of contribution/community guidelines within the software repository.
