### 2023-05-10 78c95d8
Update contribution guidelines and add code of conduct
