+++ README.md
@@ -26,7 +26,23 @@ Documentation
 
 Sarracen's documentation is hosted online at [https://sarracen.readthedocs.io](https://sarracen.readthedocs.io).
 
-Bugs / Missing Features
------------------------
+Contributing
+------------
 
-Please raise any bugs [as an issue](https://github.com/ttricco/sarracen/issues). If something does not work as you might expect, please let us know. If there are features that you feel are missing, please let us know.
+Contributions are welcomed and appreciated. Here are some ways to get involved:
+- Submitting bug reports.
+- Feature requests or suggestions.
+- Improving the documentation or providing examples.
+- Writing code to add optimizations or new features.
+
+Please use the [GitHub issue tracker](https://github.com/ttricco/sarracen/issues) to raise any bugs or to submit feature
+requests. If something does not work as you might expect, please let us know. If there are features that you feel are
+missing, please let us know.
+
+Code submissions should be submitted as a pull request. Make sure that all existing unit tests successfully pass, and 
+please add any new unit tests that are relevant. Documentation changes should also be submitted as a pull request.
+
+If you are stuck or need help, [raising an issue](https://github.com/ttricco/sarracen/issues) is a good place to start. 
+This helps us keep common issues in public view. Feel free to also [email](mailto:tstricco@mun.ca) with questions.
+
+Please note that we adhere to a [code of conduct](CODE_OF_CONDUCT.md).
