### 2023-05-10 #5263 @ttricco
@JBorrow @hlim88 We have updated our contributions guide and community code of conduct. You will find this on the GitHub front page and also on the read the docs (under "Contributing").
