### 2023-05-23 3ec5d72
vmin and vmax play nicely with log_scale
