### 2023-05-24 #5263 @hlim88
@warrickball Finish to review the software paper.

@ttricco only one minor comment is you may want to mention/compare general visualization software such as Paraview and Visit (of course these are not targeting to visualize SPH but it is well-known and widely used, also used for SPH data too).

Other than that, I think this is ready to be out! Thanks @ttricco
