+++ docs/paper/paper.bib
@@ -146,7 +146,23 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@BOOK{paraview,
+        title = {The paraview guide: a parallel visualization application},
+       author = {{Ayachit}, Utkarsh},
+         year = {2015},
+          doi = {10.5555/2789330},
+    publisher = {Kitware, Inc.}
+}
 
+@INCOLLECTION{visit,
+       author = {{Childs}, Hank and {Brugger}, Eric and {Whitlock}, Brad and {Meredith}, Jeremy and {Ahern}, Sean and {Pugmire}, David and {Biagas}, Kathleen and {Miller}, Mark and {Harrison}, Cyrus and {Weber}, Gunther H. and {Krishnan}, Hari and {Fogal}, Thomas and {Sanderson}, Allen and {Garth}, Christoph and {Bethel}, E. Wes and {Camp}, David and {R\"{u}bel}, Oliver and {Durant}, Marc and {Favre}, Jean M. and {Navr\'{a}til}, Paul},
+          doi = {10.1201/b12985},
+        title = {VisIt: An End-User Tool For Visualizing and Analyzing Very Large Data},
+    booktitle = {High Performance Visualization--Enabling Extreme-Scale Scientific Insight},
+        pages = {357-372},
+        month = {October},
+         year = {2012}
+}
 
 @ARTICLE{swiftsimio,
        author = {{Borrow}, Josh and {Borrisov}, Alexei},
