+++ docs/paper/paper.bib
@@ -150,7 +150,6 @@ archivePrefix = {arXiv},
         title = {The paraview guide: a parallel visualization application},
        author = {{Ayachit}, Utkarsh},
          year = {2015},
-          doi = {10.5555/2789330},
     publisher = {Kitware, Inc.}
 }
 
