### 2023-06-06 #5263 @ttricco
I've fixed this invalid doi. I think I lifted the doi from the publisher's site, but it doesn't seem to be a valid universal doi (evidently) and no other paper that I can find has a doi listed for it. So I have just removed it.
