### 2023-06-06 #5263 @warrickball
Thanks for taking a look too. I've been following up in the background too and arrived at the same conclusion (i.e., just remove the mysteriously bad one). I'll try to follow up on wrapping this submission up this week.
