### 2023-06-13 #5263 @warrickball
Thanks, @ttricco. I notice that the first author in the Zenodo archive is "AndrewHarris709", rather than Andrew Harris as in the JOSS paper. Can you update this in the Zenodo archive?
