### 2023-06-14 #5263 @warrickball
Thanks! Sorry for the back and forth, but one more thing: could you update the license in the Zenodo archive too? Zenodo doesn't detect all possible LICENCE files.

Along the way, I noticed the same issue affected my own JOSS submission about a year ago, which I've now fixed too. :upside_down_face:
