### 2023-06-14 #5263 @ttricco
I noticed that issue last night too when I was fixing the author names. The licence should be correct now.
